# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "安然画室 · 课程预约小程序 API",
    "version": "1.0.0",
    "description": "线下课程介绍+预约微信小程序后端接口文档。包含用户端公开接口、用户端需登录接口、管理后台接口三部分。\n\n**认证方式**: JWT Bearer Token（Header: `Authorization: Bearer <token>`）\n\n**管理后台**: 用户名密码登录，使用 nickmame 作为用户名，默认密码为环境变量 `ADMIN_PASSWORD`（默认 `admin123`）\n\n**响应格式**: 统一为 `{ code: 0, message: \"ok\", data: {} }`\n\n**开发测试账号**: openid=dev_test_openid，无需真实微信code即可登录"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "本地开发环境"
    },
    {
      "url": "https://api.anran-studio.com/api",
      "description": "生产环境"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Token，登录后获取。用户端使用微信code登录，管理端使用用户名密码登录。"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "错误码"
          },
          "message": {
            "type": "string",
            "description": "错误描述"
          },
          "data": {
            "type": "null",
            "description": "始终为 null"
          }
        }
      },
...
      "Location": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string",
            "description": "地点名称"
          },
          "address": {
            "type": "string",
            "description": "详细地址"
          },
          "latitude": {
            "type": "number",
            "description": "纬度"
          },
          "longitude": {
            "type": "number",
            "description": "经度"
          },
          "contact": {
            "type": "string",
            "description": "联系电话"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Notice": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "content": {
            "type": "string",
            "description": "公告内容"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "description": "有效期起始"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "description": "有效期截止"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive"
            ],
            "description": "状态"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "未登录或 Token 无效",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "noToken": {
                "value": {
                  "code": 401,
                  "message": "请先登录",
                  "data": null
                }
              },
              "expired": {
                "value": {
                  "code": 401,
                  "message": "登录已过期，请重新登录",
                  "data": null
                }
              },
              "invalid": {
                "value": {
                  "code": 401,
                  "message": "无效的登录凭证",
                  "data": null
                }
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "无管理员权限",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "code": 403,
              "message": "无管理员权限",
              "data": null
            }
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/admin/locations": {
      "get": {
        "tags": [
          "管理端-地点"
        ],
        "summary": "地点列表",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "全部地点列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "ok"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Location"
                      },
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "管理端-地点"
        ],
        "summary": "创建地点",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "address"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "地点名称",
                    "example": "万达校区"
                  },
                  "address": {
                    "type": "string",
                    "description": "详细地址",
                    "example": "XX市万达广场3楼301室"
                  },
                  "latitude": {
                    "type": "number",
                    "description": "纬度",
                    "default": 0,
                    "example": 30.57226
                  },
                  "longitude": {
                    "type": "number",
                    "description": "经度",
                    "default": 0,
                    "example": 104.06654
                  },
                  "contact": {
                    "type": "string",
                    "description": "联系电话",
                    "default": "",
                    "example": "13800138000"
                  }
                }
              },
              "example": {
                "name": "万达校区",
                "address": "XX市万达广场3楼301室",
                "latitude": 30.57226,
                "longitude": 104.06654,
                "contact": "13800138000"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "example": {
                  "code": 0,
                  "message": "创建成功",
                  "data": {
                    "id": 1,
                    "name": "万达校区",
                    "address": "XX市万达广场3楼301室"
                  }
                }
              }
            }
          },
          "400": {
            "description": "名称和地址不能为空",
            "content": {
              "application/json": {
                "example": {
                  "code": 400,
                  "message": "名称和地址不能为空",
                  "data": null
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/admin/locations/{id}": {
      "put": {
        "tags": [
          "管理端-地点"
        ],
        "summary": "编辑地点",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "地点ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "address": {
                    "type": "string"
                  },
                  "latitude": {
                    "type": "number"
                  },
                  "longitude": {
                    "type": "number"
                  },
                  "contact": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "name": "万达校区（新）",
                "contact": "13800138001"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "example": {
                  "code": 0,
                  "message": "更新成功",
                  "data": {
                    "id": 1,
                    "name": "万达校区（新）"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "description": "地点不存在",
            "content": {
              "application/json": {
                "example": {
                  "code": 404,
                  "message": "地点不存在",
                  "data": null
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "管理端-地点"
        ],
        "summary": "删除地点",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "地点ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "example": {
                  "code": 0,
                  "message": "删除成功",
                  "data": null
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "nullable": true,
                      "description": "操作类接口成功时返回 null"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/admin/notices": {
      "get": {
        "tags": [
          "管理端-公告"
        ],
        "summary": "公告列表",
        "description": "获取所有公告（含非启用状态），按创建时间倒序排列。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "全部公告列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "ok"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Notice"
                      },
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "管理端-公告"
        ],
        "summary": "创建公告",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "content",
                  "start_date",
                  "end_date"
                ],
                "properties": {
                  "content": {
                    "type": "string",
                    "description": "公告内容",
                    "example": "五一假期排课调整..."
                  },
                  "start_date": {
                    "type": "string",
                    "format": "date",
                    "description": "有效期起始",
                    "example": "2026-05-01"
                  },
                  "end_date": {
                    "type": "string",
                    "format": "date",
                    "description": "有效期截止",
                    "example": "2026-05-31"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "inactive"
                    ],
                    "description": "状态",
                    "default": "active",
                    "example": "active"
                  }
                }
              },
              "example": {
                "content": "五一假期排课调整，详情查看公告",
                "start_date": "2026-05-01",
                "end_date": "2026-05-06",
                "status": "active"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "example": {
                  "code": 0,
                  "message": "创建成功",
                  "data": {
                    "id": 1,
                    "content": "五一假期排课调整...",
                    "start_date": "2026-05-01",
                    "end_date": "2026-05-06",
                    "status": "active"
                  }
                }
              }
            }
          },
          "400": {
            "description": "内容、开始日期和结束日期不能为空",
            "content": {
              "application/json": {
                "example": {
                  "code": 400,
                  "message": "内容、开始日期和结束日期不能为空",
                  "data": null
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/admin/notices/{id}": {
      "put": {
        "tags": [
          "管理端-公告"
        ],
        "summary": "编辑公告",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "公告ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": {
                    "type": "string"
                  },
                  "start_date": {
                    "type": "string",
                    "format": "date"
                  },
                  "end_date": {
                    "type": "string",
                    "format": "date"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "inactive"
                    ]
                  }
                }
              },
              "example": {
                "content": "五一假期排课调整，部分课程顺延",
                "end_date": "2026-05-07"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "example": {
                  "code": 0,
                  "message": "更新成功",
                  "data": {
                    "id": 1,
                    "content": "五一假期排课调整，部分课程顺延"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "description": "公告不存在",
            "content": {
              "application/json": {
                "example": {
                  "code": 404,
                  "message": "公告不存在",
                  "data": null
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "管理端-公告"
        ],
        "summary": "删除公告",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "公告ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "example": {
                  "code": 0,
                  "message": "删除成功",
                  "data": null
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "nullable": true,
                      "description": "操作类接口成功时返回 null"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
  }
}